chore: Changes per Copilot comments
